requirements.txt
bcrypt==2.0.0
cffi==1.5.2
Django==1.7.11
django-appconf==1.0.1
django-crispy-forms==1.6.0
django-mailer==1.1
django-messages==0.5.1
django-notification==1.3.3
django-registration==2.0.4
django-registration-redux==1.4
django-user-accounts==1.3.1
django-widget-tweaks==1.4.1
lockfile==0.12.2
pbr==1.8.1
Pillow==3.1.0
pycparser==2.14
pytz==2016.1
six==1.10.0
stevedore==1.11.0
virtualenv==14.0.6
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.24.0
